The CRUD Interface
I placed the Javascript and the CSS inside index.html because I wanted to keep it one file for this review. If I had more time, I would have used authentication better and made the UI better as well. I decided not to use any libraries as it stated we were limited to the ones on the README.